requirements.txt
torch>=1.7.0
torchvision>=0.8.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
seaborn>=0.11.0
pandas>=1.1.0
tqdm>=4.50.0
pillow>=8.0.0
numpy>=1.19.0
requests>=2.25.0
